HACK: add pop elements test
The hack is that the slice wrapper stuff is added in index.js (it
shouldn't be), and the benchmark was testing impls/ stuff directly, so
was testing unwrapped versions. And in the unwrapped versions, pop is
O(n) so the benchmark was unusably slow to run.

So I hacked around this by making it test the "default impl"
(index.js) first, but this also fucks up the other tests because
index.js monkey-patches the bit trie impl. So then the bit trie impl
runs just as fast when it shouldn't.

I dunno. This business of index.js monkey-patching to add the slice
wrapper is wrong. Find a better way to mix that in to each impl.